Adding a tests directory.
Extending the '--update' of semantics to allow the specification of 
this option without further configuration options.  This now reads in a
configuration file that is writen while doing the initial install.

Resolving update semantics.  The new tests were used to confirm that
the appropriate update behavior occurs.

Documentaton updates to highlight the fact that vpy_install and other
bootstrap scripts have a limited dependency on setuptools.