A rework of the vpy_install logic to work with virtualenv 1.11.
virtualenv no longer bundles pip or setuptools,
which is a problem because we need to support
offline builds.

This change bundles wheel files for 
pip, setuptools and virtualenv.  These are 
stored in a zip file, which is included in the
vpy_install.py file.

I have confirmed that this works for Python 2.x
and 3.x.